pyproject.toml
[tool.poetry]
name = "scorecard-ai"
version = "0.2.1"
description = ""
readme = "README.md"
license = "Apache-2.0"
authors = []
packages = [
    { include = "scorecard", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.21.2"
pydantic = ">= 1.9.2, < 2.5.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"

[tool.poetry.dev-dependencies]
mypy = "0.971"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
